+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.11.1] - 2026-09-15
 - Fixed a load-time crash (`game.i18n is undefined`) that prevented Maestro settings from registering
+- Settings now register on `init` even if a later Maestro feature fails to start
 
 ## [0.11.0] - 2026-09-15
 > Foundry VTT v14 compatibility